+++ code/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
 	"log"
 	"net/http"
 	"os"
@@ -17,12 +18,24 @@ import (
 
 var errorLogger = log.New(os.Stderr, "ERROR ", log.Llongfile)
 
+type request struct {
+	AssumedRoleARN string `json:"assumed_role_arn"`
+	TokenDuration  int64  `json:"token_duration"`
+	ExpiryWindow   int64  `json:"expiry_window"`
+}
+
 type Credential struct {
-	AccessKeyId     string    `json:"access_key_id"`
-	SecretAccessKey string    `json:"secret_access_key"`
-	SessionToken    string    `json:"session_token"`
-	Expiry          time.Time `json:"expire_at"`
-	ProviderName    string    `json:"provider_name"`
+	AccessKeyId     string    `json:"AWS_ACCESS_KEY_ID"`
+	SecretAccessKey string    `json:"AWS_SECRET_ACCESS_KEY"`
+	SessionToken    string    `json:"AWS_SESSION_TOKEN"`
+	Expiry          time.Time `json:"AWS_SESSION_TOKEN_EXPIRES_AT"`
+}
+
+func clientError(status int) (events.APIGatewayProxyResponse, error) {
+	return events.APIGatewayProxyResponse{
+		StatusCode: status,
+		Body:       http.StatusText(status),
+	}, nil
 }
 
 func serverError(err error, message string) (events.APIGatewayProxyResponse, error) {
@@ -35,21 +48,44 @@ func serverError(err error, message string) (events.APIGatewayProxyResponse, err
 }
 
 func serveRequest(req events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
-	var errMsg error
-	roleARN := os.Getenv("ASSUMED_ROLE_ARN")
+	if req.HTTPMethod != "POST" {
+		return clientError(http.StatusMethodNotAllowed)
+	}
+
+	if req.Headers["Content-Type"] != "application/json" {
+		return clientError(http.StatusNotAcceptable)
+	}
+	que := new(request)
+	err := json.Unmarshal([]byte(req.Body), que)
+	if err != nil {
+		return clientError(http.StatusUnprocessableEntity)
+	}
+
+	roleARN := que.AssumedRoleARN
+	tokenDuration := que.TokenDuration
+	expiryWindow := que.ExpiryWindow
+
 	sess := session.Must(session.NewSession())
 	conf := aws.Config{}
+
+	if tokenDuration == 0 {
+		tokenDuration = 3600
+	}
+
 	if roleARN != "" {
 		var creds *credentials.Credentials
-		creds = stscreds.NewCredentials(sess, roleARN, func(p *stscreds.AssumeRoleProvider) {})
+		creds = stscreds.NewCredentials(sess, roleARN, func(p *stscreds.AssumeRoleProvider) {
+			p.Duration = time.Duration(tokenDuration) * time.Second
+			p.ExpiryWindow = time.Duration(expiryWindow)
+		})
 		conf.Credentials = creds
 	} else {
-		return serverError(errMsg, "Assumed role ARN is not set")
+		return serverError(errors.New("Assumed role ARN is not set"), "Assumed role ARN is not set")
 	}
 
 	creds, err := conf.Credentials.Get()
 	if err != nil {
-		return serverError(err, "Unable to get credential")
+		return serverError(err, "Unable to get credentials")
 	}
 
 	exp, err := conf.Credentials.ExpiresAt()
@@ -63,7 +99,6 @@ func serveRequest(req events.APIGatewayProxyRequest) (events.APIGatewayProxyResp
 		SecretAccessKey: creds.SecretAccessKey,
 		SessionToken:    creds.SessionToken,
 		Expiry:          exp,
-		ProviderName:    creds.ProviderName,
 	}
 
 	response, err := json.Marshal(cr)