+++ data.tf
@@ -66,7 +66,7 @@ data "aws_iam_policy_document" "lambda_policy" {
     ]
 
     resources = [
-      "${var.assumed_role_arn}",
+      "*",
     ]
   }
 }